@@ -1,4 +1,4 @@
-// swift-tools-version: 5.6
+// swift-tools-version: 5.9
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -6,7 +6,7 @@ import PackageDescription
 let package = Package(
     name: "flwr",
     platforms: [
-        .iOS(.v13),
+        .iOS(.v16),
         ],
     products: [
         // Products define the executables and libraries a package produces, and make them visible to other packages.
@@ -19,8 +19,8 @@ let package = Package(
         // .package(url: /* package url */, from: "1.0.0"),
         .package(url: "https://github.com/pvieito/PythonKit.git", branch: "master"),
         .package(url: "https://github.com/kewlbear/NumPy-iOS.git", branch: "main"),
-        .package(url: "https://github.com/grpc/grpc-swift.git", from: "1.0.0"),
-        .package(url: "https://github.com/apple/swift-docc-plugin", from: "1.1.0"),
+        .package(url: "https://github.com/grpc/grpc-swift.git", from: "1.22.0"),
+        .package(url: "https://github.com/apple/swift-protobuf.git", from: "1.26.0"),
     ],
     targets: [
         // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -30,7 +30,8 @@ let package = Package(
             dependencies: [
                 .product(name: "GRPC", package: "grpc-swift"),
                 .product(name: "NumPy-iOS", package: "NumPy-iOS"),
-                .product(name: "PythonKit", package: "PythonKit")],
+                .product(name: "PythonKit", package: "PythonKit"),
+                .product(name: "SwiftProtobuf", package: "swift-protobuf")],
             path: "Sources/Flower"),
         .testTarget(
             name: "FlowerTests",
