@@ -1,24 +1,9 @@
 //
 // DO NOT EDIT.
+// swift-format-ignore-file
 //
 // Generated by the protocol buffer compiler.
-// Source: transport.proto
-//
-
-//
-// Copyright 2018, gRPC Authors All rights reserved.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+// Source: flwr/proto/transport.proto
 //
 import GRPC
 import NIO
@@ -64,10 +49,8 @@ extension Flwr_Proto_FlowerServiceClientProtocol {
   }
 }
 
-#if compiler(>=5.6)
 @available(*, deprecated)
 extension Flwr_Proto_FlowerServiceClient: @unchecked Sendable {}
-#endif // compiler(>=5.6)
 
 @available(*, deprecated, renamed: "Flwr_Proto_FlowerServiceNIOClient")
 internal final class Flwr_Proto_FlowerServiceClient: Flwr_Proto_FlowerServiceClientProtocol {
@@ -123,7 +106,6 @@ internal struct Flwr_Proto_FlowerServiceNIOClient: Flwr_Proto_FlowerServiceClien
   }
 }
 
-#if compiler(>=5.6)
 @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
 internal protocol Flwr_Proto_FlowerServiceAsyncClientProtocol: GRPCClient {
   static var serviceDescriptor: GRPCServiceDescriptor { get }
@@ -199,9 +181,7 @@ internal struct Flwr_Proto_FlowerServiceAsyncClient: Flwr_Proto_FlowerServiceAsy
   }
 }
 
-#endif // compiler(>=5.6)
-
-internal protocol Flwr_Proto_FlowerServiceClientInterceptorFactoryProtocol: GRPCSendable {
+internal protocol Flwr_Proto_FlowerServiceClientInterceptorFactoryProtocol: Sendable {
 
   /// - Returns: Interceptors to use when invoking 'join'.
   func makeJoinInterceptors() -> [ClientInterceptor<Flwr_Proto_ClientMessage, Flwr_Proto_ServerMessage>]
@@ -259,15 +239,13 @@ extension Flwr_Proto_FlowerServiceProvider {
   }
 }
 
-#if compiler(>=5.6)
-
 /// To implement a server, implement an object which conforms to this protocol.
 @available(macOS 10.15, iOS 13, tvOS 13, watchOS 6, *)
-internal protocol Flwr_Proto_FlowerServiceAsyncProvider: CallHandlerProvider {
+internal protocol Flwr_Proto_FlowerServiceAsyncProvider: CallHandlerProvider, Sendable {
   static var serviceDescriptor: GRPCServiceDescriptor { get }
   var interceptors: Flwr_Proto_FlowerServiceServerInterceptorFactoryProtocol? { get }
 
-  @Sendable func join(
+  func join(
     requestStream: GRPCAsyncRequestStream<Flwr_Proto_ClientMessage>,
     responseStream: GRPCAsyncResponseStreamWriter<Flwr_Proto_ServerMessage>,
     context: GRPCAsyncServerCallContext
@@ -299,7 +277,7 @@ extension Flwr_Proto_FlowerServiceAsyncProvider {
         requestDeserializer: ProtobufDeserializer<Flwr_Proto_ClientMessage>(),
         responseSerializer: ProtobufSerializer<Flwr_Proto_ServerMessage>(),
         interceptors: self.interceptors?.makeJoinInterceptors() ?? [],
-        wrapping: self.join(requestStream:responseStream:context:)
+        wrapping: { try await self.join(requestStream: $0, responseStream: $1, context: $2) }
       )
 
     default:
@@ -308,9 +286,7 @@ extension Flwr_Proto_FlowerServiceAsyncProvider {
   }
 }
 
-#endif // compiler(>=5.6)
-
-internal protocol Flwr_Proto_FlowerServiceServerInterceptorFactoryProtocol {
+internal protocol Flwr_Proto_FlowerServiceServerInterceptorFactoryProtocol: Sendable {
 
   /// - Returns: Interceptors to use when handling 'join'.
   ///   Defaults to calling `self.makeInterceptors()`.
