@@ -2,7 +2,7 @@
 // swift-format-ignore-file
 //
 // Generated by the Swift generator plugin for the protocol buffer compiler.
-// Source: transport.proto
+// Source: flwr/proto/transport.proto
 //
 // For information on using the generated types, please see the documentation:
 //   https://github.com/apple/swift-protobuf/
@@ -76,7 +76,7 @@ enum Flwr_Proto_Code: SwiftProtobuf.Enum {
 
 extension Flwr_Proto_Code: CaseIterable {
   // The compiler won't synthesize support with the UNRECOGNIZED case.
-  static var allCases: [Flwr_Proto_Code] = [
+  static let allCases: [Flwr_Proto_Code] = [
     .ok,
     .getPropertiesNotImplemented,
     .getParametersNotImplemented,
@@ -128,7 +128,7 @@ enum Flwr_Proto_Reason: SwiftProtobuf.Enum {
 
 extension Flwr_Proto_Reason: CaseIterable {
   // The compiler won't synthesize support with the UNRECOGNIZED case.
-  static var allCases: [Flwr_Proto_Reason] = [
+  static let allCases: [Flwr_Proto_Reason] = [
     .unknown,
     .reconnect,
     .powerDisconnected,
